// src/types/supabase.d.ts
export type Database = {
  public: {
    Tables: {
// ...
      teams: {
        Row: {
          blocked_reason: string | null
          created_at: string
          email: string
          id: string
          is_banned: boolean
          is_blocked: boolean
          name: string
          tier: string
        }
        Insert: {
          blocked_reason?: string | null
          created_at?: string
          email: string
          id?: string
          is_banned?: boolean
          is_blocked?: boolean
          name: string
          tier: string
        }
        Update: {
          blocked_reason?: string | null
          created_at?: string
          email?: string
          id?: string
          is_banned?: boolean
          is_blocked?: boolean
          name?: string
          tier?: string
        }
        Relationships: [
          {
            foreignKeyName: "teams_tiers_teams"
            columns: ["tier"]
            isOneToOne: false
            referencedRelation: "tiers"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      users_teams: {
        Row: {
          added_by: string | null
          id: number
          is_default: boolean
          team_id: string
          user_id: string
        }
        Insert: {
          added_by?: string | null
          id?: number
          is_default?: boolean
          team_id: string
          user_id: string
        }
        Update: {
          added_by?: string | null
          id?: number
          is_default?: boolean
          team_id?: string
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "users_teams_added_by_user"
            columns: ["added_by"]
            isOneToOne: false
            referencedRelation: "auth_users"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "users_teams_teams_teams"
            columns: ["team_id"]
            isOneToOne: false
            referencedRelation: "teams"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "users_teams_users_users"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "auth_users"
            referencedColumns: ["id"]
          },
        ]
      }
    }
// ...
  }
}